Add files via upload
A list comprising of 234567 English words is given. Words in the list have 1 to 24 characters.  Words that
have 6 or less characters are referred to as small words. For example, the word STATS has 5 characters
and is a small word.
The objective of this project is to estimate the number of small words in the list.